apps/api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
croniter==6.0.0
cryptography==46.0.5
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.8.1
PyJWT==2.11.0
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.0.1
requests==2.32.5
rq==2.7.0
six==1.17.0
SQLAlchemy==2.0.47
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
